IOET-PHOTOGALLERY-ML/face-api/requirements.txt
setuptools>=68.0.0
wheel
flask==2.3.3
flask-cors==3.0.10
numpy==1.24.4
requests==2.31.0
gunicorn==20.1.0
pillow==9.5.0
